fetch useritems on index
