+++ pages/index.tsx
@@ -7,7 +7,7 @@ import Header from "../components/Header/Header";
 import ItemDrawer from "../components/ItemDrawer/ItemDrawer";
 import Map from "../components/map";
 import { getMapData } from "../utils/getMapData";
-import { MapData, Feature } from "../utils/types";
+import { MapData, Feature, Item } from "../utils/types";
 import { Spinner } from "../components/Spinner/Spinner";
 import addMarkers from "../utils/addMarkers";
 import { useMapStore } from "../stores/mapStore";
@@ -18,138 +18,158 @@ import { NextRouter, useRouter } from "next/router";
 import Search from "../components/Search/Search";
 import Button from "../components/Button/Button";
 import { useItemStore } from "../stores/itemStore";
+import { getUserItems } from "../utils/getUserItems";
+import { useRequestStore } from "../stores/requestStore";
 
 mapboxgl.accessToken =
-    "pk.eyJ1IjoiYXJvbjE4IiwiYSI6ImNsMzRibG9xYjB3ZjUzaW13d2s3bzVjcGkifQ.QGlBNyR336mJ2rFfFprAPg";
+  "pk.eyJ1IjoiYXJvbjE4IiwiYSI6ImNsMzRibG9xYjB3ZjUzaW13d2s3bzVjcGkifQ.QGlBNyR336mJ2rFfFprAPg";
 
 interface Session {
-    data: string;
-    status: string;
+  data: string;
+  status: string;
 }
 
 const Home: NextPage = () => {
-    const mapContainer = useRef<HTMLDivElement>(null);
-    const [zoom, setZoom] = useState(14);
-    const [mapData, setMapData] = useState<MapData | null>(null);
-    const [initialMapData, setInitialMapData] = useState<MapData | null>(null);
-    const { location } = useLocationStore();
-    const { mapRef } = useMapStore();
-    const [selectedFilter, setSelectedFilter] = useState<string>("");
-    const { setItems } = useItemStore();
-    // const [data, setData] = useState<MapData | null>(null);
+  const mapContainer = useRef<HTMLDivElement>(null);
+  const [zoom, setZoom] = useState(14);
+  const [mapData, setMapData] = useState<MapData | null>(null);
+  const [listData, setListData] = useState<Item[]>([]);
+  const [initialMapData, setInitialMapData] = useState<MapData | null>(null);
+  const { location } = useLocationStore();
+  const { mapRef } = useMapStore();
+  const [selectedFilter, setSelectedFilter] = useState<string>("");
+  const { setItems } = useItemStore();
+  const { setRequests } = useRequestStore();
+  const session = useSession();
+  // const [data, setData] = useState<MapData | null>(null);
+  let userId = session.data.user.id;
 
-    async function getAllMapData() {
-        const mapDataFetch = await getMapData();
-        setMapData(mapDataFetch);
-        setInitialMapData(mapDataFetch);
-        setItems(mapDataFetch);
+  let sum = 0;
+
+  function getRequests() {
+    for (let i = 0; i < listData.length; i++) {
+      sum = sum + listData[i].requests.length;
     }
+    return sum;
+  }
+  async function getUserData() {
+    const userItemFetch = await getUserItems(userId);
+    setListData(userItemFetch);
+    getRequests();
+    console.log(sum);
+  }
 
-    useEffect(() => {
-        getAllMapData();
-    }, []);
+  async function getAllMapData() {
+    const mapDataFetch = await getMapData();
+    setMapData(mapDataFetch);
+    setInitialMapData(mapDataFetch);
+    setItems(mapDataFetch);
+  }
 
-    const { marker } = useMarkerStore();
+  useEffect(() => {
+    getUserData();
+    getAllMapData();
+  }, []);
 
-    function resetAndSetMarkers(updatedMapData: MapData, router: NextRouter) {
-        const popUps = document.getElementsByClassName("mapboxgl-popup");
-        if (popUps[0]) popUps[0].remove();
-        marker?.forEach((m) => m.remove());
-        const markerElements = document.getElementsByClassName("marker");
-        while (markerElements.length > 0) {
-            markerElements[0].remove();
-        }
-        addMarkers(location, mapRef, updatedMapData as MapData);
+  const { marker } = useMarkerStore();
+
+  function resetAndSetMarkers(updatedMapData: MapData, router: NextRouter) {
+    const popUps = document.getElementsByClassName("mapboxgl-popup");
+    if (popUps[0]) popUps[0].remove();
+    marker?.forEach((m) => m.remove());
+    const markerElements = document.getElementsByClassName("marker");
+    while (markerElements.length > 0) {
+      markerElements[0].remove();
     }
+    addMarkers(location, mapRef, updatedMapData as MapData);
+  }
 
-    const filterMarkers = (
-        event: React.MouseEvent<HTMLButtonElement>,
-        router: NextRouter
-    ) => {
-        if (!initialMapData) {
-            return;
-        } else if ((event.target as HTMLButtonElement).value === "Free") {
-            if (selectedFilter === "Free") {
-                setSelectedFilter("");
-                setMapData(initialMapData);
-                resetAndSetMarkers(initialMapData, router);
-            } else {
-                const filteredMarkersArr: Feature[] =
-                    initialMapData?.features.filter(
-                        (feature) => feature.type === "FREE"
-                    );
-                setSelectedFilter("Free");
-                const updatedMapData: MapData = {
-                    ...initialMapData,
-                    features: filteredMarkersArr,
-                };
-                setMapData(() => updatedMapData);
-                resetAndSetMarkers(updatedMapData, router);
-            }
-        } else {
-            if (selectedFilter === "Swap") {
-                setSelectedFilter("");
-                setMapData(initialMapData);
-                resetAndSetMarkers(initialMapData, router);
-            } else {
-                const filteredMarkersArr: Feature[] =
-                    initialMapData?.features.filter(
-                        (feature) => feature.type === "SWAP"
-                    );
-                setSelectedFilter("Swap");
-                const updatedMapData: MapData = {
-                    ...initialMapData,
-                    features: filteredMarkersArr,
-                };
-                setMapData(() => updatedMapData);
-                resetAndSetMarkers(updatedMapData, router);
-            }
-            console.log(mapRef, location);
-        }
-    };
-    const [showMap, setShowMap] = useState<boolean>(true);
-    function showMapHandler() {
-        setShowMap((prev) => !prev);
-        console.log(showMap);
+  const filterMarkers = (
+    event: React.MouseEvent<HTMLButtonElement>,
+    router: NextRouter
+  ) => {
+    if (!initialMapData) {
+      return;
+    } else if ((event.target as HTMLButtonElement).value === "Free") {
+      if (selectedFilter === "Free") {
+        setSelectedFilter("");
+        setMapData(initialMapData);
+        resetAndSetMarkers(initialMapData, router);
+      } else {
+        const filteredMarkersArr: Feature[] = initialMapData?.features.filter(
+          (feature) => feature.type === "FREE"
+        );
+        setSelectedFilter("Free");
+        const updatedMapData: MapData = {
+          ...initialMapData,
+          features: filteredMarkersArr,
+        };
+        setMapData(() => updatedMapData);
+        resetAndSetMarkers(updatedMapData, router);
+      }
+    } else {
+      if (selectedFilter === "Swap") {
+        setSelectedFilter("");
+        setMapData(initialMapData);
+        resetAndSetMarkers(initialMapData, router);
+      } else {
+        const filteredMarkersArr: Feature[] = initialMapData?.features.filter(
+          (feature) => feature.type === "SWAP"
+        );
+        setSelectedFilter("Swap");
+        const updatedMapData: MapData = {
+          ...initialMapData,
+          features: filteredMarkersArr,
+        };
+        setMapData(() => updatedMapData);
+        resetAndSetMarkers(updatedMapData, router);
+      }
+      console.log(mapRef, location);
     }
+  };
+  const [showMap, setShowMap] = useState<boolean>(true);
+  function showMapHandler() {
+    setShowMap((prev) => !prev);
+    console.log(showMap);
+  }
 
-    const { data: session, status } = useSession();
+  const { data: session, status } = useSession();
 
-    const router = useRouter();
-    useEffect(() => {
-        console.log(session, status);
-        if (status === "unauthenticated") {
-            router.push("/signin");
-        }
-    }, [status]);
-
-    const loading = status === "loading";
-
-    if (loading) {
-        return <p>Loading...</p>;
+  const router = useRouter();
+  useEffect(() => {
+    console.log(session, status);
+    if (status === "unauthenticated") {
+      router.push("/signin");
     }
+  }, [status]);
 
-    return (
-        <div className="space-y-2 h-[calc(100vh-64px)]">
-            <Search properties={mapData?.features!} />
-            <div className="flex gap-2 px-2">
-                <Button
-                    type="button"
-                    selected={selectedFilter === "Free"}
-                    onClick={(evt) => filterMarkers(evt, router)}
-                    value={"Free"}
-                />
-                <Button
-                    type="button"
-                    selected={selectedFilter === "Swap"}
-                    onClick={(evt) => filterMarkers(evt, router)}
-                    value={"Swap"}
-                />
-            </div>
-            {!mapData ? <Spinner /> : <Map mapData={mapData} />}
-            <ItemDrawer selectedFilter={selectedFilter}></ItemDrawer>
-        </div>
-    );
+  const loading = status === "loading";
+
+  if (loading) {
+    return <p>Loading...</p>;
+  }
+
+  return (
+    <div className="space-y-2 h-[calc(100vh-64px)]">
+      <Search properties={mapData?.features!} />
+      <div className="flex gap-2 px-2">
+        <Button
+          type="button"
+          selected={selectedFilter === "Free"}
+          onClick={(evt) => filterMarkers(evt, router)}
+          value={"Free"}
+        />
+        <Button
+          type="button"
+          selected={selectedFilter === "Swap"}
+          onClick={(evt) => filterMarkers(evt, router)}
+          value={"Swap"}
+        />
+      </div>
+      {!mapData ? <Spinner /> : <Map mapData={mapData} />}
+      <ItemDrawer selectedFilter={selectedFilter}></ItemDrawer>
+    </div>
+  );
 };
 
 export default Home;
