+++ pages/useritems.tsx
@@ -17,25 +17,13 @@ const UserItems = (feature: Feature) => {
   const [selectedFilter, setSelectedFilter] = useState<string>("");
   const [itemDeleted, setItemDeleted] = useState<number>(0);
   const session = useSession();
-  const { setRequests } = useRequestStore();
-  let pendingReq: number;
-  let sum = 0;
-
-  function getRequests() {
-    for (let i = 0; i < listData.length; i++) {
-      sum = sum + listData[i].requests.length;
-    }
-    return sum;
-  }
-
   let userId = session.data.user.id;
   let itemId: string;
+
   async function getData() {
     const userItemFetch = await getUserItems(userId);
     setInitialUserItem(userItemFetch);
     setListData(userItemFetch);
-    getRequests();
-    console.log(sum);
   }
 
   async function deleteUserListItem(identifier: string) {
